+++ backend/.env
@@ -1,7 +1,8 @@
 DB_URL=mongodb+srv://inclusihire:MC426_inclusihire@cluster0.qmmz34t.mongodb.net/testy
 
-#Comente a linha 4 e descomente a 5 para testar usando mocha
-DB_NAME='inclusihire'
-#DB_NAME=inclusihire_test
+# Comente a linha 4 e descomente a 5 para testar usando mocha
+#DB_NAME='inclusihire'
+DB_NAME=inclusihire_test
+
 
 PORT=8080
