+++ backend/scr/controllers/controller_cadastro.js
@@ -1,0 +1,24 @@
+const modelo = require('../models/model_cadastro');
+
+async function empresaCadastro(req, res){
+    console.log(req.body);
+
+    const result = await modelo.insereNovoUsuario(req.body.nome, req.body.documento,
+         req.body.email, req.body.senha, req.body.confirmarSenha, 1);
+    
+    return res.status(200).json(result);       
+}
+
+async function usuarioCadastro(req, res){
+    console.log(req.body);
+
+    const result = await modelo.insereNovoUsuario(req.body.nome, req.body.documento,
+         req.body.email, req.body.senha, req.body.confirmarSenha, 2);
+    
+    return res.status(200).json(result);       
+}
+
+module.exports = {
+    empresaCadastro,
+    usuarioCadastro
+  };
