+++ backend/scr/routes/empresa_route.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const empresaController = require('../controllers/controller_empresa'); 
+const empresaController = require('../controllers/controller_cadastro'); 
 
 router.post('/cadastro', empresaController.empresaCadastro);
