+++ inclusihire/.gitignore
@@ -1,6 +1,7 @@
 .DS_Store
 node_modules
 /dist
+dist
 
 
 # local env files
